move modules config into service pkg
